Loader name change and Code Refactor
- set isNLSearchEnabled default value to false
- changed NLLoader to NLSearchLoader
