+++ src/atomicui/atoms/NLSearchLoader/index.ts
@@ -1,4 +1,4 @@
 /* Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved. */
 /* SPDX-License-Identifier: MIT-0 */
 
-export { default as NLLoader } from "./NLLoader";
+export { default as NLSearchLoader } from "./NLSearchLoader";
