+++ src/atomicui/organisms/AuthGeofenceBox/AuthGeofenceBox.tsx
@@ -126,8 +126,7 @@ const AuthGeofenceBox: React.FC<AuthGeofenceBoxProps> = ({ mapRef, setShowAuthGe
 						exact,
 						sg => setSuggestions(sg),
 						TriggeredByEnum.GEOFENCE_MODULE,
-						AnalyticsEventActionsEnum.AUTOCOMPLETE,
-						false
+						AnalyticsEventActionsEnum.AUTOCOMPLETE
 					);
 				}, 200);
 			}
