+++ src/atomicui/organisms/RouteBox/RouteBox.tsx
@@ -268,8 +268,7 @@ const RouteBox: React.FC<RouteBoxProps> = ({ mapRef, setShowRouteBox, isSideMenu
 								: setSuggestions({ ...suggestions, to: sg });
 						},
 						TriggeredByEnum.ROUTE_MODULE,
-						action,
-						false
+						action
 					);
 				}, 200);
 			}
